projects/03-ai-infrastructure-diagram-generator/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.4.0
httpx==0.27.0

# AI / LangChain
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.2.4
openai==1.30.1

# Terraform parsing
python-hcl2==4.3.2

# Diagram generation
graphviz==0.20.3
diagrams==0.23.4

# Vector store
chromadb==0.5.5

# AWS SDK
boto3==1.34.113
botocore==1.34.113

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
